concatenate program members
this is a temporary location as the classes should live in the _moderngl helper and imported from the c++ code
