--- moderngl/program_members/__init__.py
@@ -1,5 +1,0 @@
-from .attribute import *  # noqa
-from .subroutine import *  # noqa
-from .uniform import *  # noqa
-from .uniform_block import *  # noqa
-from .varying import *  # noqa
